requirements.txt
# Core ML/DL libraries
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers ecosystem
transformers==4.35.2
accelerate==0.25.0
datasets==2.15.0
huggingface_hub==0.19.4
trl==0.7.4
peft==0.7.1

# Quantization & optimization
bitsandbytes==0.41.3
sentencepiece==0.1.99

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
wandb==0.16.0
hf_xet==1.0.0

# Data processing
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1

# Optional: for better monitoring
ipython==8.17.2

# Web server
flask==3.0.0
flask-cors==4.0.0
